move the workorder history deletion logic to the Oban worker
